Remove uuid module
Fixes: https://github.com/parse-community/Parse-SDK-JS/issues/939
