+++ package.json
@@ -31,7 +31,6 @@
   "dependencies": {
     "@babel/runtime": "7.6.3",
     "@babel/runtime-corejs3": "7.6.3",
-    "uuid": "3.3.3",
     "ws": "7.1.2",
     "xmlhttprequest": "1.8.0"
   },
