+++ src/__tests__/ParseQuery-test.js
@@ -20,11 +20,8 @@ jest.dontMock('../ObjectStateMutations');
 jest.dontMock('../LocalDatastore');
 jest.dontMock('../OfflineQuery');
 jest.dontMock('../LiveQuerySubscription');
+jest.dontMock('../uuid');
 
-jest.mock('uuid/v4', () => {
-  let value = 0;
-  return () => value++;
-});
 const mockObject = function(className) {
   this.className = className;
   this.attributes = {};
